Refactor the validation mechanism
